Cli benchmark subcommands

+++ cli/medperf/entities/registration.py
@@ -1,17 +1,18 @@
+import os
 import yaml
 from pathlib import Path
-import os
 
-from medperf.ui import UI
 from medperf.utils import (
     get_folder_sha1,
     pretty_error,
     approval_prompt,
     dict_pretty_print,
 )
-from medperf.comms import Comms
+from medperf.ui.interface import UI
 import medperf.config as config
-from medperf.entities import Cube, Dataset
+from medperf.comms.interface import Comms
+from medperf.entities.cube import Cube
+from medperf.entities.dataset import Dataset
 
 
 class Registration:
@@ -102,13 +103,6 @@ class Registration:
 
         return registration
 
-    def retrieve_additional_data(self, ui: UI):
-        """Prompts the user for the name, description and location
-        """
-        self.name = ui.prompt("Provide a dataset name: ")
-        self.description = ui.prompt("Provide a description:  ")
-        self.location = ui.prompt("Provide a location:     ")
-
     def request_approval(self, ui: UI) -> bool:
         """Prompts the user for approval concerning uploading the registration to the comms.
 
@@ -140,7 +134,8 @@ class Registration:
         """
         uid = self.generated_uid
         new_path = os.path.join(str(Path(out_path).parent), str(uid))
-        os.rename(out_path, new_path)
+        if not os.path.exists(new_path):
+            os.rename(out_path, new_path)
         self.path = new_path
         return new_path
